+++ src/Models/Taxon.php
@@ -16,6 +16,7 @@ use Cviebrock\EloquentSluggable\SluggableScopeHelpers;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Query\Builder;
 use Illuminate\Support\Collection;
 use Vanilo\Category\Contracts\Taxon as TaxonContract;
 use Vanilo\Category\Contracts\Taxonomy as TaxonomyContract;
@@ -30,6 +31,39 @@ class Taxon extends Model implements TaxonContract
 
     /** @var Collection */
     private $_parents;
+
+    public static function findOneByParentsAndSlug(
+        string $taxonomySlug,
+        string $slug,
+        ?string $parentSlug = null
+    ): ?TaxonContract {
+        $taxonomyClass = TaxonomyProxy::modelClass();
+        $taxonomyTableName = (new $taxonomyClass())->getTable();
+        $taxonTableName = (new static())->getTable();
+        $query = static::query()
+                       ->where(function(Builder $query) use ($taxonomyTableName) {
+                           $query->select('slug')
+                               ->from($taxonomyTableName)
+                               ->whereColumn('taxonomy_id', "$taxonomyTableName.id");
+                       }, $taxonomySlug)
+                       ->where('slug', $slug);
+
+        if (null !== $parentSlug) {
+            $query->where('parent_id', function(Builder $query) use ($taxonomyTableName, $taxonomySlug, $taxonTableName, $parentSlug) {
+                $query->select('id')
+                    ->from($taxonTableName)
+                    ->where(function(Builder $query) use ($taxonomyTableName) {
+                        $query->select('slug')
+                              ->from($taxonomyTableName)
+                              ->whereColumn('taxonomy_id', "$taxonomyTableName.id");
+                    }, $taxonomySlug)
+                    ->where('slug', $parentSlug);
+            });
+        }
+
+        //return dd($query->toSql());
+        return $query->first();
+    }
 
     public function getParentsAttribute(): Collection
     {